:recycle: Refactor to components
