+++ src/components/Footer.js
@@ -2,7 +2,7 @@ import React from "react";
 import Link from "@material-ui/core/Link";
 import Typography from "@material-ui/core/Typography";
 
-export default function Copyright() {
+export default () => {
   return (
     <>
       {/* Copyright */}
